+++ db/updates/update_6.0-6.1.sql
@@ -1,0 +1,1 @@
+ALTER TABLE `allergens` ADD `toDeclare` INT NOT NULL DEFAULT '0' AFTER `percentage`;