+++ pages/listIngredients.php
@@ -19,7 +19,7 @@ $defCatClass = $settings['defCatClass'];
      </div>
      <div class="col-sm-6 text-right">
       <div class="btn-group">
-       <button type="button" class="btn btn-primary dropdown-toggle" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false"><i class="fa fa-bars"></i> Menu</button>
+       <button type="button" class="btn btn-primary dropdown-toggle" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false"><i class="fa fa-bars"></i> Actions</button>
       <div class="dropdown-menu dropdown-menu-right">
         <a class="dropdown-item popup-link" href="/pages/mgmIngredient.php">Add new ingredient</a>
         <a class="dropdown-item" id="csv_export" href="/pages/export.php?format=csv&kind=ingredients">Export to CSV</a>