+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # CHANGELOG
+### Version 6.1
+- Stock management improvements
+- Improved database and system core upgrade process
+- Add allergen materials from compositions to IFRA certification and when printing a warning label
+- Various other improvements and bug fixes
+
 ### Version 6.0
 - Redirect to the previous page if session is expired
 - Fixes errors with PHP8